==> 00_C57BL6J_2DG_WGCNA_cleaning_SM_SI_spleen.Rmd <==
---
title: "Data Cleaning Skeletal Muscle, Small Intestine, and Spleen"
author: Ann Wells
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output:
 html_document:
   keep_md: TRUE
   code_folding: hide
   collapsed: no
   toc: yes
   toc_float: yes
   self_contained: no
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

needed.packages <- c("tidyverse", "here", "functional", "gplots", "dplyr", "GeneOverlap", "R.utils", "reshape2","magrittr","data.table", "RColorBrewer","preprocessCore", "ARTool","emmeans", "phia", "gProfileR", "WGCNA")
for(i in 1:length(needed.packages)){library(needed.packages[i], character.only = TRUE)}

source(here("source_files","WGCNA_source.R"))
```

```{r load_data}
tdata.FPKM.sample.info <- readRDS(here("Data","20190406_RNAseq_B6_4wk_2DG_counts_phenotypes.RData"))

tdata.FPKM <- readRDS(here("Data","20190406_RNAseq_B6_4wk_2DG_counts_numeric.RData"))

log.tdata.FPKM <- log(tdata.FPKM + 1)
log.tdata.FPKM <- as.data.frame(log.tdata.FPKM)

log.tdata.FPKM.sample.info <- cbind(log.tdata.FPKM, tdata.FPKM.sample.info[,27238:27240])

log.tdata.FPKM.sample.info <- log.tdata.FPKM.sample.info %>% rownames_to_column() %>% filter(rowname != "A113") %>% column_to_rownames()

log.tdata.FPKM.subset <- log.tdata.FPKM[,colMeans(log.tdata.FPKM != 0) > 0.5] 

log.tdata.FPKM.sample.info.subset <- cbind(log.tdata.FPKM.subset,tdata.FPKM.sample.info[,27238:27240])
log.tdata.FPKM.sample.info.subset <- log.tdata.FPKM.sample.info.subset %>% rownames_to_column() %>% filter(rowname != "A113") %>% column_to_rownames()

log.tdata.FPKM.sample.info.subset.SM.SI.spleen <- log.tdata.FPKM.sample.info.subset %>% rownames_to_column() %>% filter(Tissue %in% c("Skeletal Muscle", "Spleen", "Small Intestine")) %>% column_to_rownames()
```

Wrangle Data
------------
I will use Mouse IDs, Tissue, Treatment, and Time to keep track of the values in the matrices. All other covariates will be discarded.

```{r}
# Set rownames by mouse ID and tissue
rownames(log.tdata.FPKM.sample.info.subset.SM.SI.spleen) <- paste0(rownames(log.tdata.FPKM.sample.info.subset.SM.SI.spleen),":", log.tdata.FPKM.sample.info.subset.SM.SI.spleen$Time, ":", log.tdata.FPKM.sample.info.subset.SM.SI.spleen$Treatment, ":", log.tdata.FPKM.sample.info.subset.SM.SI.spleen$Tissue)

# Discard covariates from columns 17333-17336
log.tdata.FPKM.sample.info.subset.SM.SI.spleen <- log.tdata.FPKM.sample.info.subset.SM.SI.spleen[,-(17333:17335)]

head(log.tdata.FPKM.sample.info.subset.SM.SI.spleen[,1:5])
```

Check Data for Missing Values
-----------------------------

WGCNA will have poor results if the data have too many missing values. I checked if any metabolites fall into this category.

```{r}
log.tdata.FPKM.sample.info.subset.SM.SI.spleen.missing <- missing(log.tdata.FPKM.sample.info.subset.SM.SI.spleen)
cat("logFPKM: ", goodSamplesGenes(log.tdata.FPKM.sample.info.subset.SM.SI.spleen.missing, verbose=0)$allOK, "\n")

```

WGCNA reports that all data are good! I now use hierarchical clustering to detect any obvious outliers. I did not see any particularly egregious outliers.

```{r, cache=TRUE}
sampleclustering(log.tdata.FPKM.sample.info.subset.SM.SI.spleen.missing)
```

```{r save_data}
saveRDS(log.tdata.FPKM.sample.info.subset.SM.SI.spleen.missing, here("Data","SM.SI.spleen","log.tdata.FPKM.sample.info.subset.SM.SI.spleen.missing.WGCNA.RData"))
```
