==> 03_C57BL6J_2DG_WGCNA_contribution.Rmd <==
---
title: "Contribution of Main Effects"
author: Ann Wells
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output:
 html_document:
   keep_md: TRUE
   code_folding: hide
   collapsed: no
   toc: yes
   toc_float: yes
   self_contained: no
   
---

# Introduction and Data files

```{r setup, echo=FALSE, warning=FALSE, message=FALSE}
options(stringsAsFactors = F)
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
knitr::opts_knit$set(progress=FALSE)
```

This dataset contains nine tissues (heart, hippocampus, hypothalamus, kidney, liver, prefrontal cortex, skeletal muscle, small intestine, and spleen) from C57BL/6J mice that were fed 2-deoxyglucose (6g/L) through their drinking water for 96hrs or 4wks. 96hr mice were given their 2DG treatment 2 weeks after the other cohort started the 4 week treatment. The organs from the mice were harvested and processed for metabolomics and transcriptomics. The data in this document pertains to the transcriptomics data only. The counts that were used were FPKM normalized before being log transformed. It was determined that sample A113 had low RNAseq quality and through further analyses with PCA, MA plots, and clustering was an outlier and will be removed for the rest of the analyses performed. This document will determine the contribution of each main effect or their combination to each module identified, as well as the relationship of each gene within the module and its significance.

```{r load packages}
needed.packages <- c("tidyverse", "here", "functional", "gplots", "dplyr", "GeneOverlap", "R.utils", "reshape2","magrittr","data.table", "RColorBrewer","preprocessCore", "ARTool","emmeans", "phia", "gProfileR", "WGCNA","plotly", "pheatmap", "ppcor","pander","downloadthis")
for(i in 1:length(needed.packages)){library(needed.packages[i], character.only = TRUE)}

source(here("source_files","WGCNA_source.R"))
source(here("source_files","WGCNA_contribution_source.R"))
```

```{r}
tdata.FPKM.sample.info <- readRDS(here("Data","20190406_RNAseq_B6_4wk_2DG_counts_phenotypes.RData"))

tdata.FPKM <- readRDS(here("Data","20190406_RNAseq_B6_4wk_2DG_counts_numeric.RData"))

log.tdata.FPKM <- log(tdata.FPKM + 1)
log.tdata.FPKM <- as.data.frame(log.tdata.FPKM)

log.tdata.FPKM.sample.info <- cbind(log.tdata.FPKM, tdata.FPKM.sample.info[,27238:27240])

log.tdata.FPKM.sample.info <- log.tdata.FPKM.sample.info %>% rownames_to_column() %>% filter(rowname != "A113") %>% column_to_rownames()

log.tdata.FPKM.subset <- log.tdata.FPKM[,colMeans(log.tdata.FPKM != 0) > 0.5] 

log.tdata.FPKM.subset <- log.tdata.FPKM.subset %>% rownames_to_column() %>% filter(rowname != "A113") %>% column_to_rownames()

WGCNA.pathway <-readRDS(here("Data","Chang_B6_96hr_4wk_gprofiler_pathway_annotation_list_WGCNA.RData"))

Matched<-readRDS(here("Data","Annotated_genes_in_WGCNA_Chang_B6_96hr_4wk.RData"))

module.names <- Matched$X..Module.
name <- str_split(module.names,"_")
samples <-c()
for(i in 1:length(name)){
samples[[i]] <- name[[i]][2]
}  
name <- str_split(samples,"\"")
name <- unlist(name)

Treatment <- unclass(as.factor(log.tdata.FPKM.sample.info[,27238]))
Tissue <- unclass(as.factor(log.tdata.FPKM.sample.info[,27239]))
Time <- unclass(as.factor(log.tdata.FPKM.sample.info[,27237]))
Treat.Tissue <- paste0(Treatment,Tissue)
Treat.Time <- paste0(Treatment, Time)
Tissue.Time <- paste0(Tissue, Time)
Treat.Tissue.Time <- paste0(Treatment,Tissue, Time)
phenotype <- data.frame(cbind(Tissue, Treatment, Time, Treat.Tissue, Treat.Time, Tissue.Time, Treat.Tissue.Time))

nSamples <- nrow(log.tdata.FPKM.sample.info)

MEs0 <- read.csv(here("Data","log.tdata.FPKM.sample.info.subset.WGCNA.module.eigens.csv"),header = T, row.names = 1)
name <- str_split(names(MEs0),"_")
samples <-c()
for(i in 1:length(name)){
samples[[i]] <- name[[i]][2]
}  
name <- str_split(samples,"\"")
name <- unlist(name)
colnames(MEs0) <-name
MEs <- orderMEs(MEs0)
moduleTraitCor <- cor(MEs, phenotype, use = "p");
moduleTraitPvalue <- corPvalueStudent(moduleTraitCor, nSamples)
```

# Relationship between Modules and Traits
```{r, results='asis', cache=TRUE}
#sizeGrWindow(10,6)
# Will display correlations and their p-values
textMatrix = paste(signif(moduleTraitCor, 2), "\n(",
                        signif(moduleTraitPvalue, 1), ")", sep = "");
dim(textMatrix) = dim(moduleTraitCor)

# Display the correlation values within a heatmap plot
heat <- pheatmap(moduleTraitCor, main = paste("Module-trait relationships"), color=colorRampPalette(brewer.pal(n = 12, name = "Paired"))(10), cluster_rows = F, cluster_cols = F, fontsize_number = 4, angle_col = 45, number_color = "black", border_color = "white")
heat
```

```{r, results='asis'}
DT::datatable(moduleTraitPvalue, extensions = 'Buttons',
                  rownames = TRUE, 
                  filter="top",
                  options = list(dom = 'Blfrtip',
                                 buttons = c('copy', 'csv', 'excel'),
                                 lengthMenu = list(c(10,25,50,-1),
                                                   c(10,25,50,"All")), 
                                 scrollX= TRUE), class = "display")
```

# Partial Correlation {.tabset .tabset-fade .tabset-pills}

```{r, results='asis'}
phenotype$Tissue <- as.numeric(phenotype$Tissue)
phenotype$Treatment <- as.numeric(phenotype$Treatment)
phenotype$Time <- as.numeric(phenotype$Time)
phenotype$Treat.Tissue <- as.numeric(phenotype$Treat.Tissue)
phenotype$Treat.Time <- as.numeric(phenotype$Treat.Time)
phenotype$Tissue.Time <- as.numeric(phenotype$Tissue.Time)
phenotype$Treat.Tissue.Time <- as.numeric(phenotype$Treat.Tissue.Time)

Sigmodules <- as.data.frame(cbind(MEs$floralwhite,MEs$darkgreen,MEs$darkolivegreen,MEs$darkorange2,MEs$lightcyan1,MEs$lightyellow,MEs$midnightblue,MEs$orangered4,MEs$royalblue,MEs$skyblue)) #MEs$ivory,MEs$skyblue3,MEs$turquoise,MEs$violet # These have a computationally singular error need to fix
colnames(Sigmodules) <- c("floralwhite","darkgreen","darkolivegreen","darkorange2","lightcyan1","lightyellow","midnightblue","orangered4","royalblue","skyblue") #"ivory","skyblue3","turquoise","violet"

for(i in 1:length(Sigmodules)){
   cat("\n##",colnames(Sigmodules[i]),"{.tabset .tabset-fade .tabset-pills}","\n")
   for(j in 1:length(phenotype)){
      cat("\n###", colnames(phenotype[j]), "{.tabset .tabset-fade .tabset-pills}", "\n")
      for(k in 1:length(phenotype)){
         cat("\n####", "Partial Correlation", colnames(phenotype[k]),"\n")
         partial <- pcor.test(Sigmodules[,i], phenotype[,j],phenotype[,k])
         panderOptions('knitr.auto.asis', FALSE)
         print(pander(partial))
         cat("\n \n")
      }
      cat("\n \n")
   }
   cat("\n \n")
}
```

# Scatterplots Module membership vs. Gene significance {.tabset .tabset-fade .tabset-pills}

```{r}
Treatment <- unclass(as.factor(log.tdata.FPKM.sample.info[,27238]))
Tissue <- unclass(as.factor(log.tdata.FPKM.sample.info[,27239]))
Time <- unclass(as.factor(log.tdata.FPKM.sample.info[,27237]))
Treat.Tissue <- paste0(Treatment,Tissue)
Treat.Time <- paste0(Treatment, Time)
Tissue.Time <- paste0(Tissue, Time)
Treat.Tissue.Time <- paste0(Treatment,Tissue, Time)
phenotype <- data.frame(cbind(Tissue, Treatment, Time, Treat.Tissue, Treat.Time, Tissue.Time, Treat.Tissue.Time))
```

## Tissue {.tabset .tabset-fade .tabset-pills}
```{r, results='asis'}
tissue.contribution(phenotype, log.tdata.FPKM.subset, MEs)
```

## Treatment {.tabset .tabset-fade .tabset-pills}
```{r, results='asis'}
treat.contribution(phenotype, log.tdata.FPKM.subset, MEs)
```

## Time {.tabset .tabset-fade .tabset-pills}
```{r, results='asis'}
time.contribution(phenotype, log.tdata.FPKM.subset, MEs)
```

## Treatment by Time {.tabset .tabset-fade .tabset-pills}
```{r, results='asis'}
treat.time.contribution(phenotype, log.tdata.FPKM.subset, MEs)
```

## Treatment by Tissue {.tabset .tabset-fade .tabset-pills}
```{r, results='asis'}
treat.tissue.contribution(phenotype, log.tdata.FPKM.subset, MEs)
```

## Time by Tissue {.tabset .tabset-fade .tabset-pills}
```{r, results='asis'}
tissue.time.contribution(phenotype, log.tdata.FPKM.subset, MEs)
```

## Treatment by Time by Tissue {.tabset .tabset-fade .tabset-pills}
```{r, results='asis'}
time.treat.tissue.contribution(phenotype, log.tdata.FPKM.subset, MEs)
```

```{r, include=FALSE}
p <- plot_ly(data = iris, x = ~Sepal.Length, y = ~Petal.Length)
p
```
